requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
SQLAlchemy>=1.4.47
alembic>=1.9.4
PyMySQL>=1.0.3
pydantic>=1.10.12,<2.0.0
cryptography>=3.4.8
celery>=5.4.0
redis>=4.5.4
python-dotenv>=1.0.0
email-validator>=1.2.1
python-multipart>=0.0.5
httpx>=0.24.0
starlette>=0.29.1